Include nexus3_resources_test in kitchen suites & cleanup tests
Now that resources are validated in nexus3_resources_test, most of the
test recipes are useless.
We should only keep the ones that are used to validate specific edge
cases, or demonstrate how to use resources.

Same thing for resources chefspec, but we need inspec tests to validate
to validate the behavior.